requirements.txt
asgiref==3.11.1
attrs==26.1.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dtcwt==0.14.0
fonttools==4.62.1
future==1.0.0
ImageIO==2.37.3
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
networkx==3.6.1
numpy==1.26.4
opencv-python==4.8.1.78
packaging==26.0
pillow==12.1.1
pydicom==3.0.2
PyJWT==2.12.1
pyparsing==3.3.2
pyswarms==1.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tifffile==2026.3.3
tqdm==4.67.3
tzdata==2025.3
